pyproject.toml
[tool.poetry]
name = "tianjiewong-portfolio"
version = "0.1.0"
description = ""
authors = ["Tian Jie Wong <wtj.wongtianjie@gmail.com>"]
readme = "README.md"
packages = [{include = "tianjiewong_portfolio"}]

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.2.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
